src/requirements.txt
Django==2.1.2
django-appconf==1.0.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-js-asset==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-taggit==0.22.2
markdown2==2.3.6
pilkit==2.0
Pillow==5.3.0
pkg-resources==0.0.0
pytz==2018.5
six==1.11.0
